#76 move existing qr tests into new dynamo0p3_quadrature_test.py and update for XYoZ
