+++ src/psyclone/tests/dynamo0p3_test.py
@@ -935,174 +935,6 @@ def test_field_fs():
     assert str(generated_code).find(output) != -1
 
 
-def test_field_qr():
-    ''' Tests that a call, with a set of fields requiring
-    quadrature, produces correct code. '''
-    _, invoke_info = parse(os.path.join(BASE_PATH,
-                                        "1.1_single_invoke_qr.f90"),
-                           api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3").create(invoke_info)
-    generated_code = str(psy.gen)
-    print generated_code
-    output_decls = (
-        "    SUBROUTINE invoke_0_testkern_qr_type(f1, f2, m1, a, m2, istp,"
-        " qr)\n"
-        "      USE testkern_qr, ONLY: testkern_qr_code\n"
-        "      USE mesh_mod, ONLY: mesh_type\n"
-        "      REAL(KIND=r_def), intent(in) :: a\n"
-        "      INTEGER, intent(in) :: istp\n"
-        "      TYPE(field_type), intent(inout) :: f1\n"
-        "      TYPE(field_type), intent(in) :: f2, m1, m2\n"
-        "      TYPE(quadrature_type), intent(in) :: qr\n"
-        "      INTEGER cell\n"
-        "      REAL(KIND=r_def), allocatable :: basis_w1_qr(:,:,:,:), "
-        "basis_w3_qr(:,:,:,:), diff_basis_w2_qr(:,:,:,:), "
-        "diff_basis_w3_qr(:,:,:,:)\n"
-        "      INTEGER dim_w1, dim_w3, diff_dim_w2, diff_dim_w3\n"
-        "      REAL(KIND=r_def), pointer :: wv_qr(:) => null(), "
-        "zp_qr(:) => null(), wh_qr(:) => null()\n"
-        "      REAL(KIND=r_def), pointer :: xp_qr(:,:) => null()\n"
-        "      INTEGER nqp_h_qr, nqp_v_qr\n"
-        "      INTEGER ndf_w1, undf_w1, ndf_w2, undf_w2, ndf_w3, undf_w3\n"
-        "      TYPE(mesh_type), pointer :: mesh => null()\n"
-        "      INTEGER nlayers\n"
-        "      TYPE(field_proxy_type) f1_proxy, f2_proxy, m1_proxy, m2_proxy\n"
-        "      INTEGER, pointer :: map_w2(:,:) => null(), "
-        "map_w3(:,:) => null(), map_w1(:,:) => null()\n")
-    assert output_decls in generated_code
-    output = (
-        "      !\n"
-        "      ! Initialise field and/or operator proxies\n"
-        "      !\n"
-        "      f1_proxy = f1%get_proxy()\n"
-        "      f2_proxy = f2%get_proxy()\n"
-        "      m1_proxy = m1%get_proxy()\n"
-        "      m2_proxy = m2%get_proxy()\n"
-        "      !\n"
-        "      ! Initialise number of layers\n"
-        "      !\n"
-        "      nlayers = f1_proxy%vspace%get_nlayers()\n"
-        "      !\n"
-        "      ! Create a mesh object\n"
-        "      !\n"
-        "      mesh => f1%get_mesh()\n"
-        "      !\n"
-        "      ! Look-up dofmaps for each function space\n"
-        "      !\n"
-        "      map_w2 => f2_proxy%vspace%get_whole_dofmap()\n"
-        "      map_w3 => m2_proxy%vspace%get_whole_dofmap()\n"
-        "      map_w1 => f1_proxy%vspace%get_whole_dofmap()\n"
-        "      !\n"
-        "      ! Initialise number of DoFs for w1\n"
-        "      !\n"
-        "      ndf_w1 = f1_proxy%vspace%get_ndf()\n"
-        "      undf_w1 = f1_proxy%vspace%get_undf()\n"
-        "      !\n"
-        "      ! Initialise number of DoFs for w2\n"
-        "      !\n"
-        "      ndf_w2 = f2_proxy%vspace%get_ndf()\n"
-        "      undf_w2 = f2_proxy%vspace%get_undf()\n"
-        "      !\n"
-        "      ! Initialise number of DoFs for w3\n"
-        "      !\n"
-        "      ndf_w3 = m2_proxy%vspace%get_ndf()\n"
-        "      undf_w3 = m2_proxy%vspace%get_undf()\n"
-        "      !\n"
-        "      ! Look-up quadrature variables\n"
-        "      !\n"
-        "      wv_qr => qr%get_wqp_v()\n"
-        "      xp_qr => qr%get_xqp_h()\n"
-        "      zp_qr => qr%get_xqp_v()\n"
-        "      wh_qr => qr%get_wqp_h()\n"
-        "      nqp_h_qr = qr%get_nqp_h()\n"
-        "      nqp_v_qr = qr%get_nqp_v()\n"
-        "      !\n"
-        "      ! Allocate basis arrays\n"
-        "      !\n"
-        "      dim_w1 = f1_proxy%vspace%get_dim_space()\n"
-        "      ALLOCATE (basis_w1_qr(dim_w1, ndf_w1, nqp_h_qr, nqp_v_qr))\n"
-        "      dim_w3 = m2_proxy%vspace%get_dim_space()\n"
-        "      ALLOCATE (basis_w3_qr(dim_w3, ndf_w3, nqp_h_qr, nqp_v_qr))\n"
-        "      !\n"
-        "      ! Allocate differential basis arrays\n"
-        "      !\n"
-        "      diff_dim_w2 = f2_proxy%vspace%get_dim_space_diff()\n"
-        "      ALLOCATE (diff_basis_w2_qr(diff_dim_w2, ndf_w2, nqp_h_qr, "
-        "nqp_v_qr))\n"
-        "      diff_dim_w3 = m2_proxy%vspace%get_dim_space_diff()\n"
-        "      ALLOCATE (diff_basis_w3_qr(diff_dim_w3, ndf_w3, nqp_h_qr, "
-        "nqp_v_qr))\n"
-        "      !\n"
-        "      ! Compute basis arrays\n"
-        "      !\n"
-        "      CALL f1_proxy%vspace%compute_basis_function(basis_w1_qr, "
-        "ndf_w1, nqp_h_qr, nqp_v_qr, xp_qr, zp_qr)\n"
-        "      CALL m2_proxy%vspace%compute_basis_function(basis_w3_qr, "
-        "ndf_w3, nqp_h_qr, nqp_v_qr, xp_qr, zp_qr)\n"
-        "      !\n"
-        "      ! Compute differential basis arrays\n"
-        "      !\n"
-        "      CALL f2_proxy%vspace%compute_diff_basis_function("
-        "diff_basis_w2_qr, ndf_w2, nqp_h_qr, nqp_v_qr, xp_qr, zp_qr)\n"
-        "      CALL m2_proxy%vspace%compute_diff_basis_function("
-        "diff_basis_w3_qr, ndf_w3, nqp_h_qr, nqp_v_qr, xp_qr, zp_qr)\n"
-        "      !\n"
-        "      ! Call kernels and communication routines\n"
-        "      !\n"
-        "      IF (f2_proxy%is_dirty(depth=1)) THEN\n"
-        "        CALL f2_proxy%halo_exchange(depth=1)\n"
-        "      END IF \n"
-        "      !\n"
-        "      IF (m1_proxy%is_dirty(depth=1)) THEN\n"
-        "        CALL m1_proxy%halo_exchange(depth=1)\n"
-        "      END IF \n"
-        "      !\n"
-        "      IF (m2_proxy%is_dirty(depth=1)) THEN\n"
-        "        CALL m2_proxy%halo_exchange(depth=1)\n"
-        "      END IF \n"
-        "      !\n"
-        "      DO cell=1,mesh%get_last_halo_cell(1)\n"
-        "        !\n"
-        "        CALL testkern_qr_code(nlayers, f1_proxy%data, f2_proxy%data, "
-        "m1_proxy%data, a, m2_proxy%data, istp, ndf_w1, undf_w1, "
-        "map_w1(:,cell), basis_w1_qr, ndf_w2, undf_w2, map_w2(:,cell), "
-        "diff_basis_w2_qr, ndf_w3, undf_w3, map_w3(:,cell), basis_w3_qr, "
-        "diff_basis_w3_qr, nqp_h_qr, nqp_v_qr, wh_qr, wv_qr)\n"
-        "      END DO \n"
-        "      !\n"
-        "      ! Set halos dirty for fields modified in the above loop\n"
-        "      !\n"
-        "      CALL f1_proxy%set_dirty()\n"
-        "      !\n"
-        "      !\n"
-        "      ! Deallocate basis arrays\n"
-        "      !\n"
-        "      DEALLOCATE (diff_basis_w2_qr, basis_w1_qr, basis_w3_qr, "
-        "diff_basis_w3_qr)\n"
-        "      !\n"
-        "    END SUBROUTINE invoke_0_testkern_qr_type"
-    )
-    assert output in generated_code
-
-
-def test_field_qr_deref():
-    ''' Tests that a call, with a set of fields requiring
-    quadrature, produces correct code when the quadrature is supplied as the
-    component of a derived type. '''
-    _, invoke_info = parse(os.path.join(BASE_PATH,
-                                        "1.1.1_single_invoke_qr_deref.f90"),
-                           api="dynamo0.3")
-    for dist_mem in [True, False]:
-        psy = PSyFactory("dynamo0.3",
-                         distributed_memory=dist_mem).create(invoke_info)
-        gen = str(psy.gen)
-        print gen
-        assert (
-            "    SUBROUTINE invoke_0_testkern_qr_type(f1, f2, m1, a, m2, istp,"
-            " qr_data)\n" in gen)
-        assert "TYPE(quadrature_type), intent(in) :: qr_data" in gen
-
-
 def test_real_scalar():
     ''' tests that we generate correct code when a kernel takes a single,
     real scalar argument (plus fields)'''
